requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
dateutils==0.6.12
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
selenium==4.20.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
zope.interface==6.3
